move simple git command functions from vara-manager to new git_command module and refactor all to use plumbum
