+++ varats/varats/tools/research_tools/vara_manager.py
@@ -14,12 +14,12 @@ from enum import Enum
 from pathlib import Path
 from threading import RLock
 
-from benchbuild.utils.cmd import git, grep
-from plumbum import RETCODE, TF, local
+from benchbuild.utils.cmd import git
+from plumbum import local
 from PyQt5.QtCore import QProcess
 
 from varats.utils.exceptions import ProcessTerminatedError
-from varats.utils.git_util import get_current_branch, CommitHash
+from varats.utils.git_commands import fetch_remote
 from varats.utils.settings import vara_cfg
 
 LOG = logging.getLogger(__name__)
@@ -33,37 +33,6 @@ def run_process_with_output(
     output = str(process.readAllStandardOutput().data().decode('utf-8'))
     for line in output.splitlines(True):
         post_out(line)
-
-
-def download_repo(
-    dl_folder: Path,
-    url: str,
-    repo_name: tp.Optional[str] = None,
-    remote_name: tp.Optional[str] = None,
-    post_out: tp.Callable[[str], None] = lambda x: None
-) -> None:
-    """Download a repo into the specified folder."""
-    if not dl_folder.exists():
-        raise Exception(
-            "Could not find download folder  {dl_folder}".format(
-                dl_folder=dl_folder
-            )
-        )
-
-    with local.cwd(dl_folder):
-        args = ["clone", "--progress", url]
-        if remote_name is not None:
-            args.append("--origin")
-            args.append(remote_name)
-
-        if repo_name is not None:
-            args.append(repo_name)
-
-        with ProcessManager.create_process("git", args) as proc:
-            proc.setProcessChannelMode(QProcess.MergedChannels)
-            proc.readyReadStandardOutput.connect(
-                lambda: run_process_with_output(proc, post_out)
-            )
 
 
 class BuildType(Enum):
@@ -104,174 +73,6 @@ class BuildType(Enum):
         if suffix:
             return Path(f"{str(self)}_{suffix}")
         return Path(str(self))
-
-
-def add_remote(repo_folder: Path, remote: str, url: str) -> None:
-    """Adds new remote to the repository."""
-    with ProcessManager.create_process(
-        "git", ["remote", "add", remote, url], workdir=repo_folder
-    ):
-        pass
-
-
-def show_status(repo_folder: Path) -> None:
-    """Show git status."""
-    with local.cwd(repo_folder):
-        git["status"].run_fg()
-
-
-def get_branches(
-    repo_folder: Path, extra_args: tp.Optional[tp.List[str]] = None
-) -> str:
-    """Show git branches."""
-    extra_args = [] if extra_args is None else extra_args
-
-    args = ["branch"]
-    args += extra_args
-
-    with local.cwd(repo_folder):
-        return tp.cast(str, git(args))
-
-
-def fetch_remote(
-    remote: tp.Optional[str] = None,
-    repo_folder: tp.Optional[Path] = None,
-    extra_args: tp.Optional[tp.List[str]] = None
-) -> None:
-    """Fetches the new changes from the remote."""
-    extra_args = [] if extra_args is None else extra_args
-
-    args = ["fetch"]
-    args += extra_args
-    if remote:
-        args.append(remote)
-
-    with ProcessManager.create_process("git", args, workdir=repo_folder):
-        pass
-
-
-def get_tags(repo_folder: Path,
-             extra_args: tp.Optional[tp.List[str]] = None) -> tp.List[str]:
-    """Get the list of available git tags."""
-
-    extra_args = [] if extra_args is None else extra_args
-    args = ["tag"]
-    args += extra_args
-
-    with local.cwd(repo_folder):
-        git_tag_string: str = git(args)
-        git_tag_list: tp.List[str] = []
-
-        if git_tag_string:
-            git_tag_list = git_tag_string.split("\n")
-            git_tag_list.remove('')
-            return git_tag_list
-
-        return git_tag_list
-
-
-def init_all_submodules(folder: Path) -> None:
-    """Inits all submodules."""
-    with ProcessManager.create_process(
-        "git", ["submodule", "init"], workdir=folder
-    ):
-        pass
-
-
-def update_all_submodules(folder: Path, recursive: bool = True) -> None:
-    """Updates all submodules."""
-    git_params = ["submodule", "update"]
-    if recursive:
-        git_params.append("--recursive")
-
-    with ProcessManager.create_process("git", git_params, workdir=folder):
-        pass
-
-
-def pull_current_branch(repo_folder: tp.Optional[Path] = None) -> None:
-    """Pull in changes in a certain branch."""
-    with ProcessManager.create_process("git", ["pull"], workdir=repo_folder):
-        pass
-
-
-def push_current_branch(
-    repo_folder: tp.Optional[Path] = None,
-    upstream: tp.Optional[str] = None,
-    branch_name: tp.Optional[str] = None
-) -> None:
-    """Push in changes in a certain branch."""
-    cmd_args = ["push"]
-
-    if upstream is not None:
-        cmd_args.append("--set-upstream")
-        cmd_args.append(upstream)
-        if branch_name is not None:
-            cmd_args.append(branch_name)
-        else:
-            cmd_args.append(get_current_branch(repo_folder))
-
-    if repo_folder is None or repo_folder == Path(""):
-        git(cmd_args)
-    else:
-        with local.cwd(repo_folder):
-            git(cmd_args)
-
-
-def fetch_repository(repo_folder: tp.Optional[Path] = None) -> None:
-    """Pull in changes in a certain branch."""
-    with ProcessManager.create_process("git", ["fetch"], workdir=repo_folder):
-        pass
-
-
-def checkout_branch_or_commit(
-    repo_folder: Path, target: tp.Union[str, CommitHash]
-) -> None:
-    """Checks out a branch or commit in the repository."""
-    with ProcessManager.create_process(
-        "git", ["checkout", str(target)], workdir=repo_folder
-    ):
-        pass
-
-
-def checkout_new_branch(
-    repo_folder: Path,
-    branch: str,
-    remote_branch: tp.Optional[str] = None
-) -> None:
-    """Checks out a new branch in the repository."""
-    args = ["checkout", "-b", branch]
-    if remote_branch is not None:
-        args.append(remote_branch)
-    with ProcessManager.create_process("git", args, workdir=repo_folder):
-        pass
-
-
-def has_branch(repo_folder: Path, branch_name: str) -> bool:
-    """Checks if a branch exists in the local repository."""
-    with local.cwd(repo_folder):
-        exit_code = git["rev-parse", "--verify", branch_name] & TF
-        return tp.cast(bool, exit_code)
-
-
-def has_remote_branch(repo_folder: Path, branch_name: str, remote: str) -> bool:
-    """Checks if a remote branch of a repository exists."""
-    with local.cwd(repo_folder):
-        exit_code = (
-            git["ls-remote", "--heads", remote, branch_name] | grep[branch_name]
-        ) & RETCODE
-        return tp.cast(bool, exit_code == 0)
-
-
-def branch_has_upstream(
-    repo_folder: Path, branch_name: str, upstream: str = 'origin'
-) -> bool:
-    """Check if a branch has an upstream remote."""
-    with local.cwd(repo_folder):
-        exit_code = (
-            git["rev-parse", "--abbrev-ref", branch_name + "@{upstream}"] |
-            grep[upstream]
-        ) & RETCODE
-        return tp.cast(bool, exit_code == 0)
 
 
 ###############################################################################
